+++ src/mechanism.hpp
@@ -147,16 +147,6 @@ namespace zmq
         bool check_socket_type (const std::string& type_) const;
     };
 
-    class mechanism_base_t : public mechanism_t
-    {
-      protected:
-        mechanism_base_t (session_base_t *const session_,
-                          const options_t &options_);
-
-        session_base_t *const session;
-
-        int check_basic_command_structure (msg_t *msg_);
-    };
-    }
+}
 
 #endif
