+++ src/mechanism_base.hpp
@@ -27,53 +27,23 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#ifndef __ZMQ_PLAIN_CLIENT_HPP_INCLUDED__
-#define __ZMQ_PLAIN_CLIENT_HPP_INCLUDED__
+#ifndef __ZMQ_MECHANISM_BASE_HPP_INCLUDED__
+#define __ZMQ_MECHANISM_BASE_HPP_INCLUDED__
 
-#include "mechanism_base.hpp"
-#include "options.hpp"
+#include "mechanism.hpp"
 
 namespace zmq
 {
+class mechanism_base_t : public mechanism_t
+{
+  protected:
+    mechanism_base_t (session_base_t *const session_,
+                      const options_t &options_);
 
-    class msg_t;
+    session_base_t *const session;
 
-    class plain_client_t : public mechanism_base_t
-    {
-      public:
-        plain_client_t (session_base_t *const session_,
-                        const options_t &options_);
-        virtual ~plain_client_t ();
-
-        // mechanism implementation
-        virtual int next_handshake_command (msg_t *msg_);
-        virtual int process_handshake_command (msg_t *msg_);
-        virtual status_t status () const;
-
-      private:
-
-        enum state_t {
-            sending_hello,
-            waiting_for_welcome,
-            sending_initiate,
-            waiting_for_ready,
-            error_command_received,
-            ready
-        };
-
-        state_t state;
-
-        int produce_hello (msg_t *msg_) const;
-        int produce_initiate (msg_t *msg_) const;
-
-        int process_welcome (
-            const unsigned char *cmd_data, size_t data_size);
-        int process_ready (
-            const unsigned char *cmd_data, size_t data_size);
-        int process_error (
-            const unsigned char *cmd_data, size_t data_size);
-    };
-
+    int check_basic_command_structure (msg_t *msg_);
+};
 }
 
 #endif
